test(gql_version): strengthen integration and module import tests
This addresses two issues with gql version detection testing:

- Extracts gql version parsing/detection logic to `moneyflow/gql_version.py`. This allows `test_detect_gql_v4_plus_without_gql` to accurately simulate the `gql` module being unavailable without crashing due to the unconditional import in `monarchmoney.py`.

- Adds a lightweight instantiation of `AIOHTTPTransport` and `Client` to `test_detect_gql_version_and_api` that exercises the constructor path to prevent silent breaks on future gql releases.
